# openapi_service_catalogue_management.yaml
openapi: 3.0.3
info:
  title: Service Inventory Management for federated SMC
  description:  |
    This is the Service Inventory Management OpenAPI specification for federated SMC capability (last update: Sep 08, 2023).
  version: Spiral 5
...
servers:
- url: https://mnp.api.host:443/context/serviceInventoryManagement/v5.0
paths:
  /service:
    get:
      tags:
      - service
      summary: List or find Service objects
      description: This operation list or find Service entities
      operationId: listService
      parameters:
        - in: query
          name: filter
          description: Filter query parameter to define a query filter for retrieving only resources which fulfil the filter query based on attribute names, operators and valid attribute values
          schema:
            type: object
            additionalProperties:
              type: string
              maxLength: 128
          style: form
          explode: true
        - in: query
          name: defaultFilter
          description: A process specific default filter is always implicitly set or added to the custom filter. If no custom filter is set and the parameter is set to 'false', no default filter will be set
          schema:
            type: boolean
            default: true
        - in: query
          name: allProviders
          description: Only the records for the 'main' provider will be returned by default. the records for all providers will returned, if the parameter is set to 'true'
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service_RetrieveResponseArray'
              examples:
                exampleServiceRetrieveResponseArray:
                  $ref: '#/components/examples/exampleServiceRetrieveResponseArray'
        400:
          $ref: '#/components/responses/http400'
        404:
          $ref: '#/components/responses/http404'
        500:
          $ref: '#/components/responses/http500'        
  /service/{serviceId}:
    get:
      tags:
      - service
      summary: Retrieves a Service by ID
      description: This operation retrieves a Service entity. Attribute selection
        is enabled for all first level attributes.
      operationId: retrieveService
      parameters:
      - name: serviceId
        in: path
        description: Identifier of the Service
        required: true
        schema:
          $ref: '#/components/schemas/SvcId'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service_CreateRetrieve'
              examples:
                exampleServiceRetrieveResponse:
                  $ref: '#/components/examples/exampleServiceRetrieveResponse'
        400:
          $ref: '#/components/responses/http400'
        404:
          $ref: '#/components/responses/http404'
        500:
          $ref: '#/components/responses/http500'
  /serviceInstance:
    get:
      tags:
      - serviceInstance
      summary: List or find Service Instance objects
      description: This operation list or find Service Instance entities
      operationId: listServiceInstance
      parameters:
        - in: query
          name: filter
          description: Filter query parameter to define a query filter for retrieving only resources which fulfil the filter query based on attribute names, operators and valid attribute values
          schema:
            type: object
            additionalProperties:
              type: string
          style: form
          explode: true
        - in: query
          name: defaultFilter
          description: A process specific default filter is always implicitly set or added to the custom filter. If no custom filter is set and the parameter is set to 'false', no default filter will be set
          schema:
            type: boolean
            default: true
        - in: query
          name: allProviders
          description: Only the records for the 'main' provider will be returned by default. the records for all providers will returned, if the parameter is set to 'true'
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInstance_RetrieveResponseArray'
              examples:
                exampleServiceInstanceRetrieveResponseArray:
                  $ref: '#/components/examples/exampleServiceInstanceRetrieveResponseArray'
        400:
          $ref: '#/components/responses/http400'
        404:
          $ref: '#/components/responses/http404'
        500:
          $ref: '#/components/responses/http500'
  /serviceInstance/{serviceInstanceId}:
    get:
      tags:
      - serviceInstance
      summary: Retrieves a Service Instance by ID
      description: This operation retrieves a Service Instance entity. Attribute selection
        is enabled for all first level attributes.
      operationId: retrieveServiceInstance
      parameters:
      - name: serviceInstanceId
        in: path
        description: Identifier of the Service Instance
        required: true
        schema:
          $ref: '#/components/schemas/SviId'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInstance_CreateRetrieve'
              examples:
                exampleServiceInstanceRetrieveResponse:
                  $ref: '#/components/examples/exampleServiceInstanceRetrieveResponse'
        400:
          $ref: '#/components/responses/http400'
        404:
          $ref: '#/components/responses/http404'
        500:
          $ref: '#/components/responses/http500'
  /hub:
    post:
      tags:
      - events subscription
      summary: Register a listener
      description: Sets the communication endpoint address the service instance must use to deliver information about its health state, execution state, failures and metrics.
      operationId: registerListener
      requestBody:
        description: Data containing the callback endpoint to deliver the information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventSubscription'
            examples:
              exampleEventSubscriptionWithoutQueryRequest:
                $ref: '#/components/examples/exampleEventSubscriptionWithoutQueryRequest'
              exampleEventSubscriptionWithQueryRequest:
                $ref: '#/components/examples/exampleEventSubscriptionWithQueryRequest'
      responses:
        201:
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSubscription'
              examples:
                exampleEventSubscriptionResponse:
                  $ref: '#/components/examples/exampleEventSubscriptionResponse'
        400:
          $ref: '#/components/responses/http400'
        404:
          $ref: '#/components/responses/http404'
        500:
          $ref: '#/components/responses/http500'    
      x-codegen-request-body-name: data
  /hub/{hubId}:
    delete:
      tags:
      - events subscription
      summary: Unregister a listener
      description: Resets the communication endpoint address the service instance must use to deliver information about its health state, execution state, failures and metrics.
      operationId: unregisterListener
      parameters:
        - name: hubId
          in: path
          description: The id of the registered listener
          required: true
          schema:
            $ref: '#/components/schemas/SubId'
      responses:
        204:
          $ref: '#/components/responses/http204'
        400:
          $ref: '#/components/responses/http400'
        404:
          $ref: '#/components/responses/http404'
        500:
          $ref: '#/components/responses/http500'  
  /listener:
    post:
      operationId: notifyListener
      summary: Notifies a listener
      description: Example of a listener for receiving the notification events
      tags:
        - events notification
      requestBody:
        description: The event data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventNotificationRequest'
            examples:
              exampleEventNotificationServiceCreate:
                $ref: '#/components/examples/exampleEventNotificationServiceCreate'
              exampleEventNotificationServiceUpdate:
                $ref: '#/components/examples/exampleEventNotificationServiceUpdate'
              exampleEventNotificationServiceInstanceCreate:
                $ref: '#/components/examples/exampleEventNotificationServiceInstanceCreate'
              exampleEventNotificationServiceInstanceUpdate:
                $ref: '#/components/examples/exampleEventNotificationServiceInstanceUpdate'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventNotificationResponse'
              examples:
                exampleEventNotificationResponse:
                  $ref: '#/components/examples/exampleEventNotificationResponse'
        400:
          $ref: '#/components/responses/http400'
        404:
          $ref: '#/components/responses/http404'
        500:
          $ref: '#/components/responses/http500'
components:
  schemas:
...
    Service_CreateRetrieve:
      x-sip-conformance-table-request-column-title: Create Service Notification
      type: object
      required:
      - id
      - description
      - name
      - state
      - federatedServiceExtension
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/SvcId'
        category:
          $ref: '#/components/schemas/SvcCategory'
        description:
          $ref: '#/components/schemas/SvcDescription'
        endDate:
          $ref: '#/components/schemas/SvcEndDate'
        federatedServiceExtension:
          $ref: '#/components/schemas/FederatedServiceExtension_CreateRetrieve'
        name:
          $ref: '#/components/schemas/SvcName'
        place:
          $ref: '#/components/schemas/SvcPlace'
        serviceRelationship:
          $ref: '#/components/schemas/SvcServiceRelationship'
        startDate:
          $ref: '#/components/schemas/SvcStartDate'
        state:
          $ref: '#/components/schemas/ServiceStateType'
      description: 'Service is a base class for defining the Service hierarchy. All
        Services are characterized as either being possibly visible and usable by
        a Customer or not. This gives rise to the two subclasses of Service: CustomerFacingService
        and ResourceFacingService.'
...
    Service_RetrieveResponseArray:
      type: array
      items:
        $ref: '#/components/schemas/Service_CreateRetrieve'
    ServiceInstance_CreateRetrieve:
      x-sip-conformance-table-request-column-title: Create Service Instance Notification
      type: object
      required:
      - id
      - name
      - state
      - federatedServiceInstanceExtension
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/SviId'
        category:
          $ref: '#/components/schemas/SvcCategory'
        description:
          $ref: '#/components/schemas/SvcDescription'
        endDate:
          $ref: '#/components/schemas/SvcEndDate'
        federatedServiceInstanceExtension:
          $ref: '#/components/schemas/FederatedServiceInstanceExtension_CreateRetrieve'
        name:
          $ref: '#/components/schemas/SvcName'
        place:
          $ref: '#/components/schemas/SvcPlace'
        serviceCharacteristic:
          $ref: '#/components/schemas/SvcCharacteristic'
        serviceSpecification:
          $ref: '#/components/schemas/ServiceSpecificationRef'
        startDate:
          $ref: '#/components/schemas/SvcStartDate'
        state:
          $ref: '#/components/schemas/ServiceInstanceStateType'
        supportingResource:
          $ref: '#/components/schemas/ResourceRefArray'
      description: 'Service is a base class for defining the Service hierarchy. All
        Services are characterized as either being possibly visible and usable by
        a Customer or not. This gives rise to the two subclasses of Service: CustomerFacingService
        and ResourceFacingService.'
...
    ServiceInstance_RetrieveResponseArray:
      type: array
      items:
        $ref: '#/components/schemas/ServiceInstance_CreateRetrieve'
    ### Block with schemas existing in original TMF specification
...
    EventSubscription:
      description: Sets the communication endpoint address the service instance must use to deliver notification information    
      required:
      - callback
      - id
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/SubId'
        callback:
          description: The callback being registered.
          type: string
          maxLength: 2048
          format: uri
        query:
          description: additional data to be passed
          type: string
          maxLength: 2048
        resource:
          description: As default, a subscription is set for all resources. If a subscription should be set for only one of the resources or a subscription filter query is used, the resource must be set
          type: string
          enum:
            - Service
            - ServiceInstance  
    # SMC comment: EventSubscriptionInput removed (id required for Request and therefor equals to EventSubscription)
    # SMC comment: ChangeRequestCreateEvent, ChangeRequestStatusChangeEvent, ChangeRequestAttributeValueChangeEvent, ChangeRequestApprovalRequiredEvent, ChangeRequestDeleteEvent removed (using old style notification pattern)
...
    EventNotificationRequest:
      oneOf:
        - $ref: '#/components/schemas/EventNotificationCreateService'
        - $ref: '#/components/schemas/EventNotificationUpdateService'
        - $ref: '#/components/schemas/EventNotificationCreateServiceInstance'
        - $ref: '#/components/schemas/EventNotificationUpdateServiceInstance'
...
    EventNotificationResponse:
      additionalProperties: false
      properties: {}
    ### Block with schemas for error handling
...
    SubId:
      type: string
      description: Id of the listener
      maxLength: 64
      pattern: '^[A-Z0-9]{3}-[A-Z0-9]{3}-[A-Z0-9]{3}-SUB-[a-zA-Z0-9_-]{1,48}$'            
...
    SvcId:
      type: string
      description: Unique identifier of the service
      maxLength: 64
      pattern: '^[A-Z0-9]{3}-[A-Z0-9]{3}-[A-Z0-9]{3}-SVC-[a-zA-Z0-9_-]{1,48}$'
...
    SviId:
      type: string
      description: Unique identifier of the service instance
      maxLength: 64
      pattern: '^[A-Z0-9]{3}-[A-Z0-9]{3}-[A-Z0-9]{3}-SVI-[a-zA-Z0-9_-]{1,48}$'
...
  responses:
    http204:
      description: No Content
    http400:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            exampleError400:
              $ref: '#/components/examples/exampleError400'
    http404:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            exampleError404:
              $ref: '#/components/examples/exampleError404'
    http500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            exampleError500:
              $ref: '#/components/examples/exampleError500'
  examples:
...
    exampleServiceRetrieveResponse:
      summary: Retrieve Service by id (CIS Service)
      value:
        category: labelling
        description: Service for test reasons
        id: TST-MNP-002-SVC-BWA403
        name: Platform Labelling Service
        startDate: '2022-03-30T23:22:00.000Z'
        endDate: '2023-03-30T21:22:00.000Z'
        state: active
        place:
        - id: TST-MNP-002-GAD-12345
          role: serviceProvidingLocation
        serviceRelationship:
        - relationshipType: reliesOn
          service:
            id: TST-MNP-002-SVC-12345
        federatedServiceExtension:
          lastUpdate: '2022-03-30T21:22:00.000Z'
          poc: person@organization.org
          consumer:
          - id: TST-MNP-003
          serviceC3Taxonomy:
            c3TaxonomyVersion: '5'
            c3TaxonomyId: id-6226703b-0575-4d13-a350-a7913ad83f3b
            c3TaxonomyName: Platform CIS Security Services
          securityMarking:
            securityPolicy: NATO
            securityClassification: UNCLASSIFIED
            releasabilityCommunity: AUS,AUT,CHE,FIN,NZL,SWE,UKR,EU EEAS only
    exampleServiceRetrieveResponseArray:
      summary: List Services
      value:
        - category: labelling
          description: Service for test reasons
          id: TST-MNP-002-SVC-BWA403
          name: Platform Labelling Service
          startDate: '2022-03-30T23:22:00.000Z'
          endDate: '2023-03-30T21:22:00.000Z'
          state: active
          place:
          - id: TST-MNP-002-GAD-12345
            role: serviceProvidingLocation
          serviceRelationship:
          - relationshipType: reliesOn
            service:
              id: TST-MNP-002-SVC-12345
          federatedServiceExtension:
            lastUpdate: '2022-03-30T21:22:00.000Z'
            poc: person@organization.org
            consumer:
            - id: TST-MNP-003
            serviceC3Taxonomy:
              c3TaxonomyVersion: '5'
              c3TaxonomyId: id-6226703b-0575-4d13-a350-a7913ad83f3b
              c3TaxonomyName: Platform CIS Security Services
            securityMarking:
              securityPolicy: NATO
              securityClassification: UNCLASSIFIED
              releasabilityCommunity: AUS,AUT,CHE,FIN,NZL,SWE,UKR,EU EEAS only
        - category: itsm
          description: IT Service Management Environment
          id: TST-MNP-002-SVC-123457
          name: IT Service Management
          startDate: '2022-04-30T23:22:00.000Z'
          endDate: '2023-04-30T21:22:00.000Z'
          state: active
          place:
          - id: TST-MNP-002-GAD-12345
            role: serviceProvidingLocation
          - id: TST-MNP-003-GAD-12345
            role: serviceConsumingLocation
          serviceRelationship:
          - relationshipType: reliesOn
            service:
              id: TST-MNP-002-SVC-123456
          federatedServiceExtension:
            lastUpdate: '2022-03-29T21:22:00.000Z'
            consumer:
            - id: TST-MNP-003
            serviceC3Taxonomy:
              c3TaxonomyVersion: '5'
              c3TaxonomyId: id-d46e44a9-17df-4751-9d70-a42676e18dd3
              c3TaxonomyName: COI-Enabling Services
            securityMarking:
              securityPolicy: NATO
              securityClassification: UNCLASSIFIED
              releasabilityCommunity: AUS,AUT,CHE,FIN,NZL,SWE,UKR,EU EEAS only
    exampleServiceInstanceRetrieveResponse:
      summary: Retrieve Service Instance by id
      value:
        category: labelling
        description: Service for test reasons
        id: TST-MNP-002-SVI-BWA403-001
        name: Platform Labelling Service
        startDate: '2022-03-30T23:22:00.000Z'
        endDate: '2023-03-30T21:22:00.000Z'
        state: operational
        place:
        - id: TST-MNP-002-GAD-12345
          role: serviceProvidingLocation
        - id: TST-MNP-003-GAD-123456
          role: serviceConsumingLocation
        serviceSpecification:
          id: TST-MNP-002-SSP-ABC0123456789
        serviceCharacteristic:
        - id: TST-MNP-002-SPC-1234
          name: systemName
          valueType: string
          value: MN Labelling Instance A
        - id: TST-MNP-002-SPC-1235
          name: maximumRequestsPerSecond
          valueType: number
          value: 100
        supportingResource:
        - id: TST-MNP-003-FCI-1234
          relationshipType: accessedVia
          "@referredType": API
        federatedServiceInstanceExtension:
          instanceOf: TST-MNP-002-SVC-BAW403
          lastUpdate: '2022-03-30T21:22:00.000Z'
          plannedMaintenanceStart: '2022-04-03T21:22:00.000Z'
          plannedMaintenanceEnd: '2022-04-04T21:22:00.000Z'
          poc: person@organization.org
          consumer:
          - id: TST-MNP-003
          - id: TST-MNP-031
          securityMarking:
            securityPolicy: NATO
            securityClassification: UNCLASSIFIED
            releasabilityCommunity: AUS,AUT,CHE,FIN,NZL,SWE,UKR,EU EEAS only
    exampleServiceInstanceRetrieveResponseArray:
      summary: List Service Instances
      value:
        - category: labelling
          description: Service for test reasons
          id: TST-MNP-002-SVI-BWA403-001
          name: Platform Labelling Service
          startDate: '2022-03-30T23:22:00.000Z'
          endDate: '2023-03-30T21:22:00.000Z'
          state: operational
          place:
          - id: TST-MNP-002-GAD-12345
            role: serviceProvidingLocation
          - id: TST-MNP-003-GAD-123456
            role: serviceConsumingLocation
          serviceSpecification:
            id: TST-MNP-002-SSP-ABC0123456789
          serviceCharacteristic:
          - id: TST-MNP-002-SPC-1234
            name: systemName
            valueType: string
            value: MN Labelling Instance A
          - id: TST-MNP-002-SPC-1235
            name: maximumRequestsPerSecond
            valueType: number
            value: 100
          supportingResource:
          - id: TST-MNP-003-FCI-1234
            relationshipType: accessedVia
            "@referredType": API
          federatedServiceInstanceExtension:
            instanceOf: TST-MNP-002-SVC-BAW403
            lastUpdate: '2022-03-30T21:22:00.000Z'
            plannedMaintenanceStart: '2022-04-03T21:22:00.000Z'
            plannedMaintenanceEnd: '2022-04-04T21:22:00.000Z'
            poc: person@organization.org
            consumer:
            - id: TST-MNP-003
            - id: TST-MNP-031
            securityMarking:
              securityPolicy: NATO
              securityClassification: UNCLASSIFIED
              releasabilityCommunity: AUS,AUT,CHE,FIN,NZL,SWE,UKR,EU EEAS only
        - id: TST-MNP-002-SVI-BWA403-002
          name: Platform Labelling Service
          state: operational
          federatedServiceInstanceExtension:
            instanceOf: TST-MNP-002-SVC-BAW403
            lastUpdate: '2022-03-30T21:22:00.000Z'
            consumer:
            - id: TST-MNP-003
            securityMarking:
              securityPolicy: NATO
              securityClassification: UNCLASSIFIED
              releasabilityCommunity: AUS,AUT,CHE,FIN,NZL,SWE,UKR,EU EEAS only   
    ################ Examples for subscriptions ################
    exampleEventSubscriptionWithoutQueryRequest:
      summary: Create Subscription w/o Filter 
      value:
        id: TST-MNP-001-SUB-12345
        callback: https://mnp.api.host:443/context/serviceInventoryManagement/v5.0/listener    
    exampleEventSubscriptionWithQueryRequest:
      summary: Create Subscription with Filter (and concrete resource)
      value:
        id: TST-MNP-001-SUB-12345
        callback: https://mnp.api.host:443/context/serviceInventoryManagement/v5.0/listener
        resource: Service
        query: id.eq=TST-MNP-002-SVC-12345
    exampleEventSubscriptionResponse:
      summary: Create Subscription response
      value:
        id: TST-MNP-001-SUB-12345
        callback: https://mnp.api.host:443/context/serviceInventoryManagement/v5.0/listener       
    ################ Examples for notifications ################
    exampleEventNotificationServiceCreate:
      summary: Example for a create Service Notification
      value:
        eventType: ServiceCreateEvent
        eventTime: '2019-02-26T15:38:27.248Z'
        eventId: TST-MNP-002-PUB-1931744701538
        event:
          service:
            id: TST-MNP-002-SVC-BWA403
            description: Service for test reasons
            name: Platform Labelling Service
            state: active
            federatedServiceExtension:
              lastUpdate: '2022-03-30T21:22:00.000Z'
              serviceC3Taxonomy:
                c3TaxonomyVersion: '5'
                c3TaxonomyId: id-6226703b-0575-4d13-a350-a7913ad83f3b
                c3TaxonomyName: Platform CIS Security Services
              securityMarking:
                securityPolicy: NATO
                securityClassification: UNCLASSIFIED
                releasabilityCommunity: AUS,AUT,CHE,FIN,NZL,SWE,UKR,EU EEAS only
    exampleEventNotificationServiceUpdate:
      summary: Example for an update Service Notification
      value:
        eventType: ServiceAttributeValueChangeEvent
        eventTime: '2019-02-26T16:38:27.248Z'
        eventId: TST-MNP-002-PUB-1931744701539
        event:
          service:
            id: TST-MNP-002-SVC-BWA403
            state: inactive
            federatedServiceExtension:
              lastUpdate: '2022-03-30T21:22:00.000Z'
              securityMarking:
                securityPolicy: NATO
                securityClassification: UNCLASSIFIED
                releasabilityCommunity: AUS,AUT,CHE,FIN,NZL,SWE,UKR,EU EEAS only
    exampleEventNotificationServiceInstanceCreate:
      summary: Example for a create Service Instance Notification
      value:
        eventType: ServiceInstanceCreateEvent
        eventTime: '2019-02-26T17:38:27.248Z'
        eventId: TST-MNP-002-PUB-19317447700
        event:
          serviceInstance:
            id: TST-MNP-002-SVI-BWA403-002
            name: Platform Labelling Service
            state: operational
            federatedServiceInstanceExtension:
              instanceOf: TST-MNP-002-SVC-BAW403
              lastUpdate: '2022-03-30T21:22:00.000Z'
              consumer:
              - id: TST-MNP-003
              securityMarking:
                securityPolicy: NATO
                securityClassification: UNCLASSIFIED
                releasabilityCommunity: AUS,AUT,CHE,FIN,NZL,SWE,UKR,EU EEAS only
...
    exampleEventNotificationResponse:
      summary: Notification response
      value: {}
    ################ Examples for error responses ################      
...